refactor in-memory components of file authority to InMemoryAuthority
