+++ crates/server/src/store/file/authority.rs
@@ -7,41 +7,26 @@
 
 //! All authority related types
 
-#[cfg(feature = "dnssec")]
-use std::borrow::Borrow;
 use std::collections::BTreeMap;
+use std::ops::Deref;
 use std::path::{Path, PathBuf};
-use std::sync::Arc;
 
-use futures::future::{self, Future, FutureResult, IntoFuture};
+use futures::future::Future;
 
-use trust_dns::op::{LowerQuery, ResponseCode};
+use trust_dns::op::LowerQuery;
+use trust_dns::proto::rr::dnssec::rdata::key::KEY;
 use trust_dns::rr::dnssec::{DnsSecResult, Signer, SupportedAlgorithms};
-use trust_dns::rr::{DNSClass, LowerName, Name, RData, Record, RecordSet, RecordType, RrKey};
+use trust_dns::rr::{LowerName, Name, RecordSet, RecordType, RrKey};
 
-use authority::{
-    AnyRecords, AuthLookup, Authority, LookupError, LookupRecords, LookupResult, MessageRequest,
-    UpdateResult, ZoneType,
-};
+use authority::{Authority, LookupError, MessageRequest, UpdateResult, ZoneType};
 use store::file::FileConfig;
+use store::in_memory::InMemoryAuthority;
 
-/// SqliteAuthority is responsible for storing the resource records for a particular zone.
+/// FileAuthority is responsible for storing the resource records for a particular zone.
 ///
 /// Authorities default to DNSClass IN. The ZoneType specifies if this should be treated as the
 /// start of authority for the zone, is a slave, or a cached zone.
-pub struct FileAuthority {
-    origin: LowerName,
-    class: DNSClass,
-    records: BTreeMap<RrKey, Arc<RecordSet>>,
-    zone_type: ZoneType,
-    allow_axfr: bool,
-    // Private key mapped to the Record of the DNSKey
-    //  TODO: these private_keys should be stored securely. Ideally, we have keys only stored per
-    //   server instance, but that requires requesting updates from the parent zone, which may or
-    //   may not support dynamic updates to register the new key... Trust-DNS will provide support
-    //   for this, in some form, perhaps alternate root zones...
-    secure_keys: Vec<Signer>,
-}
+pub struct FileAuthority(InMemoryAuthority);
 
 impl FileAuthority {
     /// Creates a new Authority.
@@ -65,17 +50,9 @@ impl FileAuthority {
         zone_type: ZoneType,
         allow_axfr: bool,
     ) -> Self {
-        Self {
-            origin: LowerName::new(&origin),
-            class: DNSClass::IN,
-            records: records
-                .into_iter()
-                .map(|(key, rrset)| (key, Arc::new(rrset)))
-                .collect(),
-            zone_type,
-            allow_axfr,
-            secure_keys: Vec::new(),
-        }
+        Self(InMemoryAuthority::new(
+            origin, records, zone_type, allow_axfr,
+        ))
     }
 
     /// Read the Authority for the origin from the specified configuration
@@ -120,422 +97,42 @@ impl FileAuthority {
         Ok(FileAuthority::new(origin, records, zone_type, allow_axfr))
     }
 
-    /// Enables AXFRs of all the zones records
-    pub fn set_allow_axfr(&mut self, allow_axfr: bool) {
-        self.allow_axfr = allow_axfr;
+    /// Unwrap the InMemoryAuthority
+    pub fn unwrap(self) -> InMemoryAuthority {
+        self.0
     }
+}
 
-    /// Retrieve the Signer, which contains the private keys, for this zone
-    pub fn secure_keys(&self) -> &[Signer] {
-        &self.secure_keys
-    }
+impl Deref for FileAuthority {
+    type Target = InMemoryAuthority;
 
-    /// Get all the
-    pub fn records(&self) -> &BTreeMap<RrKey, Arc<RecordSet>> {
-        &self.records
-    }
-
-    /// Returns the minimum ttl (as used in the SOA record)
-    pub fn minimum_ttl(&self) -> u32 {
-        let soa = Authority::soa(self).wait(/*TODO: this works because the future here is always complete*/);
-
-        let soa = match soa {
-            Ok(soa) => soa,
-            Err(e) => {
-                warn!("could not lookup SOA for authority: {}: {}", self.origin, e);
-                return 0;
-            }
-        };
-
-        soa.iter().next().map_or(0, |soa| {
-            if let RData::SOA(ref rdata) = *soa.rdata() {
-                rdata.minimum()
-            } else {
-                0
-            }
-        })
-    }
-
-    /// get the current serial number for the zone.
-    pub fn serial(&self) -> u32 {
-        let soa = Authority::soa(self).wait(/*TODO: this works because the future here is always complete*/);
-
-        let soa = match soa {
-            Ok(soa) => soa,
-            Err(e) => {
-                warn!("could not lookup SOA for authority: {}: {}", self.origin, e);
-                return 0;
-            }
-        };
-
-        soa.iter().next().map_or_else(
-            || {
-                warn!("no soa record found for zone: {}", self.origin);
-                0
-            },
-            |soa| {
-                if let RData::SOA(ref soa_rdata) = *soa.rdata() {
-                    soa_rdata.serial()
-                } else {
-                    panic!("This was not an SOA record"); // valid panic, never should happen
-                }
-            },
-        )
-    }
-
-    #[allow(unused)]
-    fn increment_soa_serial(&mut self) -> u32 {
-        let soa = Authority::soa(self).wait(/*TODO: this works because the future here is always complete*/);
-
-        let soa = match soa {
-            Ok(soa) => soa,
-            Err(e) => {
-                warn!("could not lookup SOA for authority: {}: {}", self.origin, e);
-                return 0;
-            }
-        };
-
-        let opt_soa_serial = soa.iter().next().map(|soa| {
-            // TODO: can we get a mut reference to SOA directly?
-            let mut soa: Record = soa.clone();
-
-            let serial = if let RData::SOA(ref mut soa_rdata) = *soa.rdata_mut() {
-                soa_rdata.increment_serial();
-                soa_rdata.serial()
-            } else {
-                panic!("This was not an SOA record"); // valid panic, never should happen
-            };
-
-            (soa, serial)
-        });
-
-        if let Some((soa, serial)) = opt_soa_serial {
-            self.upsert(soa, serial);
-            serial
-        } else {
-            error!(
-                "no soa record found for zone while attempting increment: {}",
-                self.origin
-            );
-            0
-        }
-    }
-
-    /// Inserts or updates a `Record` depending on it's existence in the authority.
-    ///
-    /// Guarantees that SOA, CNAME only has one record, will implicitly update if they already exist.
-    ///
-    /// # Arguments
-    ///
-    /// * `record` - The `Record` to be inserted or updated.
-    /// * `serial` - Current serial number to be recorded against updates.
-    ///
-    /// # Return value
-    ///
-    /// Ok() on success or Err() with the `ResponseCode` associated with the error.
-    pub fn upsert(&mut self, record: Record, serial: u32) -> bool {
-        assert_eq!(self.class, record.dns_class());
-
-        let rr_key = RrKey::new(record.name().into(), record.rr_type());
-        let records: &mut Arc<RecordSet> = self
-            .records
-            .entry(rr_key)
-            .or_insert_with(|| Arc::new(RecordSet::new(record.name(), record.rr_type(), serial)));
-
-        // because this is and Arc, we need to clone and then replace the entry
-        let mut records_clone = RecordSet::clone(&*records);
-        if records_clone.insert(record, serial) {
-            *records = Arc::new(records_clone);
-            true
-        } else {
-            false
-        }
-    }
-
-    /// (Re)generates the nsec records, increments the serial number nad signs the zone
-    #[cfg(feature = "dnssec")]
-    pub fn secure_zone(&mut self) -> DnsSecResult<()> {
-        // TODO: only call nsec_zone after adds/deletes
-        // needs to be called before incrementing the soa serial, to make sur IXFR works properly
-        self.nsec_zone();
-
-        // need to resign any records at the current serial number and bump the number.
-        // first bump the serial number on the SOA, so that it is resigned with the new serial.
-        self.increment_soa_serial();
-
-        // TODO: should we auto sign here? or maybe up a level...
-        self.sign_zone()
-    }
-
-    /// (Re)generates the nsec records, increments the serial number nad signs the zone
-    #[cfg(not(feature = "dnssec"))]
-    pub fn secure_zone(&mut self) -> Result<(), &str> {
-        Err("DNSSEC was not enabled during compilation.")
-    }
-
-    /// Dummy implementation for when DNSSEC is disabled.
-    #[cfg(feature = "dnssec")]
-    fn nsec_zone(&mut self) {
-        use trust_dns::rr::rdata::{DNSSECRData, DNSSECRecordType, NSEC};
-
-        // only create nsec records for secure zones
-        if self.secure_keys.is_empty() {
-            return;
-        }
-        debug!("generating nsec records: {}", self.origin);
-
-        // first remove all existing nsec records
-        let delete_keys: Vec<RrKey> = self
-            .records
-            .keys()
-            .filter(|k| k.record_type == RecordType::DNSSEC(DNSSECRecordType::NSEC))
-            .cloned()
-            .collect();
-
-        for key in delete_keys {
-            self.records.remove(&key);
-        }
-
-        // now go through and generate the nsec records
-        let ttl = self.minimum_ttl();
-        let serial = self.serial();
-        let mut records: Vec<Record> = vec![];
-
-        {
-            let mut nsec_info: Option<(&Name, Vec<RecordType>)> = None;
-            for key in self.records.keys() {
-                match nsec_info {
-                    None => nsec_info = Some((key.name.borrow(), vec![key.record_type])),
-                    Some((name, ref mut vec)) if LowerName::new(name) == key.name => {
-                        vec.push(key.record_type)
-                    }
-                    Some((name, vec)) => {
-                        // names aren't equal, create the NSEC record
-                        let mut record = Record::with(
-                            name.clone(),
-                            RecordType::DNSSEC(DNSSECRecordType::NSEC),
-                            ttl,
-                        );
-                        let rdata = NSEC::new_cover_self(key.name.clone().into(), vec);
-                        record.set_rdata(RData::DNSSEC(DNSSECRData::NSEC(rdata)));
-                        records.push(record);
-
-                        // new record...
-                        nsec_info = Some((&key.name.borrow(), vec![key.record_type]))
-                    }
-                }
-            }
-
-            // the last record
-            if let Some((name, vec)) = nsec_info {
-                // names aren't equal, create the NSEC record
-                let mut record = Record::with(
-                    name.clone(),
-                    RecordType::DNSSEC(DNSSECRecordType::NSEC),
-                    ttl,
-                );
-                let rdata = NSEC::new_cover_self(Authority::origin(self).clone().into(), vec);
-                record.set_rdata(RData::DNSSEC(DNSSECRData::NSEC(rdata)));
-                records.push(record);
-            }
-        }
-
-        // insert all the nsec records
-        for record in records {
-            self.upsert(record, serial);
-        }
-    }
-
-    /// Signs any records in the zone that have serial numbers greater than or equal to `serial`
-    #[cfg(feature = "dnssec")]
-    fn sign_zone(&mut self) -> DnsSecResult<()> {
-        use chrono::Utc;
-        use trust_dns::rr::dnssec::tbs;
-        use trust_dns::rr::rdata::{DNSSECRData, DNSSECRecordType, SIG};
-
-        debug!("signing zone: {}", self.origin);
-        let inception = Utc::now();
-        let zone_ttl = self.minimum_ttl();
-
-        // TODO: should this be an error?
-        if self.secure_keys.is_empty() {
-            warn!("attempt to sign_zone for dnssec, but no keys available!")
-        }
-
-        // sign all record_sets, as of 0.12.1 this includes DNSKEY
-        for rr_set_orig in self.records.values_mut() {
-            let mut rr_set = RecordSet::clone(&*rr_set_orig);
-            // becuase the rrset is an Arc, it must be cloned before mutated
-
-            rr_set.clear_rrsigs();
-            let rrsig_temp = Record::with(
-                rr_set.name().clone(),
-                RecordType::DNSSEC(DNSSECRecordType::RRSIG),
-                zone_ttl,
-            );
-
-            for signer in &self.secure_keys {
-                debug!(
-                    "signing rr_set: {}, {} with: {}",
-                    rr_set.name(),
-                    rr_set.record_type(),
-                    signer.algorithm(),
-                );
-
-                let expiration = inception + signer.sig_duration();
-
-                let tbs = tbs::rrset_tbs(
-                    rr_set.name(),
-                    self.class,
-                    rr_set.name().num_labels(),
-                    rr_set.record_type(),
-                    signer.algorithm(),
-                    rr_set.ttl(),
-                    expiration.timestamp() as u32,
-                    inception.timestamp() as u32,
-                    signer.calculate_key_tag()?,
-                    signer.signer_name(),
-                    // TODO: this is a nasty clone... the issue is that the vec
-                    //  from records is of Vec<&R>, but we really want &[R]
-                    &rr_set
-                        .records_without_rrsigs()
-                        .cloned()
-                        .collect::<Vec<Record>>(),
-                );
-
-                // TODO, maybe chain these with some ETL operations instead?
-                let tbs = match tbs {
-                    Ok(tbs) => tbs,
-                    Err(err) => {
-                        error!("could not serialize rrset to sign: {}", err);
-                        continue;
-                    }
-                };
-
-                let signature = signer.sign(&tbs);
-                let signature = match signature {
-                    Ok(signature) => signature,
-                    Err(err) => {
-                        error!("could not sign rrset: {}", err);
-                        continue;
-                    }
-                };
-
-                let mut rrsig = rrsig_temp.clone();
-                rrsig.set_rdata(RData::DNSSEC(DNSSECRData::SIG(SIG::new(
-                    // type_covered: RecordType,
-                    rr_set.record_type(),
-                    // algorithm: Algorithm,
-                    signer.algorithm(),
-                    // num_labels: u8,
-                    rr_set.name().num_labels(),
-                    // original_ttl: u32,
-                    rr_set.ttl(),
-                    // sig_expiration: u32,
-                    expiration.timestamp() as u32,
-                    // sig_inception: u32,
-                    inception.timestamp() as u32,
-                    // key_tag: u16,
-                    signer.calculate_key_tag()?,
-                    // signer_name: Name,
-                    signer.signer_name().clone(),
-                    // sig: Vec<u8>
-                    signature,
-                ))));
-
-                rr_set.insert_rrsig(rrsig);
-            }
-
-            *rr_set_orig = Arc::new(rr_set);
-        }
-
-        Ok(())
-    }
-
-    /// unwrap all the records
-    pub(crate) fn unwrap_records(self) -> BTreeMap<RrKey, Arc<RecordSet>> {
-        self.records
+    fn deref(&self) -> &Self::Target {
+        &self.0
     }
 }
 
 impl Authority for FileAuthority {
-    type Lookup = AuthLookup;
-    type LookupFuture = FutureResult<Self::Lookup, LookupError>;
+    type Lookup = <InMemoryAuthority as Authority>::Lookup;
+    type LookupFuture = <InMemoryAuthority as Authority>::LookupFuture;
 
     /// What type is this zone
     fn zone_type(&self) -> ZoneType {
-        self.zone_type
+        self.0.zone_type()
     }
 
     /// Return true if AXFR is allowed
     fn is_axfr_allowed(&self) -> bool {
-        self.allow_axfr
+        self.0.is_axfr_allowed()
     }
 
-    /// Takes the UpdateMessage, extracts the Records, and applies the changes to the record set.
-    ///
-    /// [RFC 2136](https://tools.ietf.org/html/rfc2136), DNS Update, April 1997
-    ///
-    /// ```text
-    ///
-    /// 3.4 - Process Update Section
-    ///
-    ///   Next, the Update Section is processed as follows.
-    ///
-    /// 3.4.2 - Update
-    ///
-    ///   The Update Section is parsed into RRs and these RRs are processed in
-    ///   order.
-    ///
-    /// 3.4.2.1. If any system failure (such as an out of memory condition,
-    ///   or a hardware error in persistent storage) occurs during the
-    ///   processing of this section, signal SERVFAIL to the requestor and undo
-    ///   all updates applied to the zone during this transaction.
-    ///
-    /// 3.4.2.2. Any Update RR whose CLASS is the same as ZCLASS is added to
-    ///   the zone.  In case of duplicate RDATAs (which for SOA RRs is always
-    ///   the case, and for WKS RRs is the case if the ADDRESS and PROTOCOL
-    ///   fields both match), the Zone RR is replaced by Update RR.  If the
-    ///   TYPE is SOA and there is no Zone SOA RR, or the new SOA.SERIAL is
-    ///   lower (according to [RFC1982]) than or equal to the current Zone SOA
-    ///   RR's SOA.SERIAL, the Update RR is ignored.  In the case of a CNAME
-    ///   Update RR and a non-CNAME Zone RRset or vice versa, ignore the CNAME
-    ///   Update RR, otherwise replace the CNAME Zone RR with the CNAME Update
-    ///   RR.
-    ///
-    /// 3.4.2.3. For any Update RR whose CLASS is ANY and whose TYPE is ANY,
-    ///   all Zone RRs with the same NAME are deleted, unless the NAME is the
-    ///   same as ZNAME in which case only those RRs whose TYPE is other than
-    ///   SOA or NS are deleted.  For any Update RR whose CLASS is ANY and
-    ///   whose TYPE is not ANY all Zone RRs with the same NAME and TYPE are
-    ///   deleted, unless the NAME is the same as ZNAME in which case neither
-    ///   SOA or NS RRs will be deleted.
-    ///
-    /// 3.4.2.4. For any Update RR whose class is NONE, any Zone RR whose
-    ///   NAME, TYPE, RDATA and RDLENGTH are equal to the Update RR is deleted,
-    ///   unless the NAME is the same as ZNAME and either the TYPE is SOA or
-    ///   the TYPE is NS and the matching Zone RR is the only NS remaining in
-    ///   the RRset, in which case this Update RR is ignored.
-    ///
-    /// 3.4.2.5. Signal NOERROR to the requestor.
-    /// ```
-    ///
-    /// # Arguments
-    ///
-    /// * `update` - The `UpdateMessage` records will be extracted and used to perform the update
-    ///              actions as specified in the above RFC.
-    ///
-    /// # Return value
-    ///
-    /// true if any of additions, updates or deletes were made to the zone, false otherwise. Err is
-    ///  returned in the case of bad data, etc.
-    fn update(&mut self, _update: &MessageRequest) -> UpdateResult<bool> {
-        Err(ResponseCode::NotImp)
+    /// Perform a dynamic update of a zone
+    fn update(&mut self, update: &MessageRequest) -> UpdateResult<bool> {
+        self.0.update(update)
     }
 
     /// Get the origin of this zone, i.e. example.com is the origin for www.example.com
     fn origin(&self) -> &LowerName {
-        &self.origin
+        self.0.origin()
     }
 
     /// Looks up all Resource Records matching the giving `Name` and `RecordType`.
@@ -559,108 +156,32 @@ impl Authority for FileAuthority {
         is_secure: bool,
         supported_algorithms: SupportedAlgorithms,
     ) -> Self::LookupFuture {
-        let rr_key = RrKey::new(name.clone(), rtype);
-
-        // Collect the records from each rr_set
-        let result: LookupResult<LookupRecords> = match rtype {
-            RecordType::AXFR | RecordType::ANY => {
-                let result = AnyRecords::new(
-                    is_secure,
-                    supported_algorithms,
-                    self.records.values().cloned().collect(),
-                    rtype,
-                    name.clone(),
-                );
-                Ok(LookupRecords::AnyRecords(result))
-            }
-            _ => self.records.get(&rr_key).map_or(
-                Err(LookupError::from(ResponseCode::NXDomain)),
-                |rr_set| {
-                    Ok(LookupRecords::new(
-                        is_secure,
-                        supported_algorithms,
-                        rr_set.clone(),
-                    ))
-                },
-            ),
-        };
-
-        // This is annoying. The 1035 spec literally specifies that most DNS authorities would want to store
-        //   records in a list except when there are a lot of records. But this makes indexed lookups by name+type
-        //   always return empty sets. This is only important in the negative case, where other DNS authorities
-        //   generally return NoError and no results when other types exist at the same name. bah.
-        let result = match result {
-            Err(LookupError::ResponseCode(ResponseCode::NXDomain)) => {
-                if self
-                    .records
-                    .keys()
-                    .any(|key| key.name() == name || name.zone_of(key.name()))
-                {
-                    return Err(LookupError::NameExists).into_future();
-                } else {
-                    return Err(LookupError::from(ResponseCode::NXDomain)).into_future();
-                }
-            }
-            Err(e) => return Err(e).into_future(),
-            o => o,
-        };
-
-        result.map(AuthLookup::from).into_future()
+        self.0.lookup(name, rtype, is_secure, supported_algorithms)
     }
 
-    // FIXME: move back to a suplied method in the trait
+    /// Using the specified query, perform a lookup against this zone.
+    ///
+    /// # Arguments
+    ///
+    /// * `query` - the query to perform the lookup with.
+    /// * `is_secure` - if true, then RRSIG records (if this is a secure zone) will be returned.
+    ///
+    /// # Return value
+    ///
+    /// Returns a vectory containing the results of the query, it will be empty if not found. If
+    ///  `is_secure` is true, in the case of no records found then NSEC records will be returned.
     fn search(
         &self,
         query: &LowerQuery,
         is_secure: bool,
         supported_algorithms: SupportedAlgorithms,
     ) -> Box<Future<Item = Self::Lookup, Error = LookupError> + Send> {
-        debug!("searching FileAuthority for: {}", query);
+        self.0.search(query, is_secure, supported_algorithms)
+    }
 
-        let lookup_name = query.name();
-        let record_type: RecordType = query.query_type();
-
-        // if this is an AXFR zone transfer, verify that this is either the slave or master
-        //  for AXFR the first and last record must be the SOA
-        if RecordType::AXFR == record_type {
-            // TODO: support more advanced AXFR options
-            if !self.is_axfr_allowed() {
-                return Box::new(Err(LookupError::from(ResponseCode::Refused)).into_future());
-            }
-
-            match self.zone_type() {
-                ZoneType::Master | ZoneType::Slave => (),
-                // TODO: Forward?
-                _ => return Box::new(Err(LookupError::from(ResponseCode::NXDomain)).into_future()),
-            }
-        }
-
-        // perform the actual lookup
-        match record_type {
-            RecordType::SOA => {
-                Box::new(self.lookup(self.origin(), record_type, is_secure, supported_algorithms))
-            }
-            RecordType::AXFR => {
-                // TODO: shouldn't these SOA's be secure? at least the first, perhaps not the last?
-                let lookup = self
-                    .soa_secure(is_secure, supported_algorithms)
-                    .join3(
-                        self.soa(),
-                        self.lookup(lookup_name, record_type, is_secure, supported_algorithms),
-                    )
-                    .map(|(start_soa, end_soa, records)| match start_soa {
-                        l @ AuthLookup::Empty => l,
-                        start_soa => AuthLookup::AXFR {
-                            start_soa: start_soa.unwrap_records(),
-                            records: records.unwrap_records(),
-                            end_soa: end_soa.unwrap_records(),
-                        },
-                    });
-
-                Box::new(lookup)
-            }
-            _ => Box::new(self.lookup(lookup_name, record_type, is_secure, supported_algorithms)),
-        }
+    /// Get the NS, NameServer, record for the zone
+    fn ns(&self, is_secure: bool, supported_algorithms: SupportedAlgorithms) -> Self::LookupFuture {
+        self.0.ns(is_secure, supported_algorithms)
     }
 
     /// Return the NSEC records based on the given name
@@ -670,162 +191,56 @@ impl Authority for FileAuthority {
     /// * `name` - given this name (i.e. the lookup name), return the NSEC record that is less than
     ///            this
     /// * `is_secure` - if true then it will return RRSIG records as well
-    #[cfg(feature = "dnssec")]
     fn get_nsec_records(
         &self,
         name: &LowerName,
         is_secure: bool,
         supported_algorithms: SupportedAlgorithms,
     ) -> Self::LookupFuture {
-        use trust_dns::rr::rdata::DNSSECRecordType;
-
-        fn is_nsec_rrset(rr_set: &RecordSet) -> bool {
-            rr_set.record_type() == RecordType::DNSSEC(DNSSECRecordType::NSEC)
-        }
-
-        // TODO: need a BorrowdRrKey
-        let rr_key = RrKey::new(name.clone(), RecordType::DNSSEC(DNSSECRecordType::NSEC));
-        let no_data = self
-            .records
-            .get(&rr_key)
-            .map(|rr_set| LookupRecords::new(is_secure, supported_algorithms, rr_set.clone()));
-
-        if no_data.is_some() {
-            return future::result(Ok(no_data.unwrap().into()));
-        }
-
-        let get_closest_nsec = |name: &LowerName| -> Option<Arc<RecordSet>> {
-            self.records
-                .values()
-                .filter(|rr_set| is_nsec_rrset(rr_set))
-                // the name must be greater than the name in the nsec
-                .filter(|rr_set| *name >= rr_set.name().into())
-                // now find the next record where the covered name is greater
-                .find(|rr_set| {
-                    // there should only be one record
-                    rr_set
-                        .records(false, SupportedAlgorithms::default())
-                        .next()
-                        .and_then(|r| r.rdata().as_dnssec())
-                        .and_then(|r| r.as_nsec())
-                        .map_or(false, |r| {
-                            // the search name is less than the next NSEC record
-                            *name < r.next_domain_name().into() ||
-                            // this is the last record, and wraps to the beginning of the zone
-                            r.next_domain_name() < rr_set.name()
-                        })
-                })
-                .cloned()
-        };
-
-        let closest_proof = get_closest_nsec(name);
-
-        // we need the wildcard proof, but make sure that it's still part of the zone.
-        let wildcard = name.base_name();
-        let wildcard = if self.origin().zone_of(&wildcard) {
-            wildcard
-        } else {
-            self.origin().clone()
-        };
-
-        // don't duplicate the record...
-        let wildcard_proof = if wildcard != *name {
-            get_closest_nsec(&wildcard)
-        } else {
-            None
-        };
-
-        let proofs = match (closest_proof, wildcard_proof) {
-            (Some(closest_proof), Some(wildcard_proof)) => {
-                // dedup with the wildcard proof
-                if wildcard_proof != closest_proof {
-                    vec![wildcard_proof, closest_proof]
-                } else {
-                    vec![closest_proof]
-                }
-            }
-            (None, Some(proof)) | (Some(proof), None) => vec![proof],
-            (None, None) => vec![],
-        };
-
-        future::result(Ok(LookupRecords::many(
-            is_secure,
-            supported_algorithms,
-            proofs,
-        )
-        .into()))
+        self.0
+            .get_nsec_records(name, is_secure, supported_algorithms)
     }
 
-    #[cfg(not(feature = "dnssec"))]
-    fn get_nsec_records(
+    /// Returns the SOA of the authority.
+    ///
+    /// *Note*: This will only return the SOA, if this is fullfilling a request, a standard lookup
+    ///  should be used, see `soa_secure()`, which will optionally return RRSIGs.
+    fn soa(&self) -> Self::LookupFuture {
+        self.0.soa()
+    }
+
+    /// Returns the SOA record for the zone
+    fn soa_secure(
         &self,
-        _name: &LowerName,
-        _is_secure: bool,
-        _supported_algorithms: SupportedAlgorithms,
+        is_secure: bool,
+        supported_algorithms: SupportedAlgorithms,
     ) -> Self::LookupFuture {
-        future::result(Ok(AuthLookup::default()))
+        self.0.soa_secure(is_secure, supported_algorithms)
     }
 
-    /// By adding a secure key, this will implicitly enable dnssec for the zone.
-    ///
-    /// # Arguments
-    ///
-    /// * `signer` - Signer with associated private key
-    #[cfg(feature = "dnssec")]
+    /// Add a (Sig0) key that is authorized to perform updates against this authority
+    fn add_update_auth_key(&mut self, name: Name, key: KEY) -> DnsSecResult<()> {
+        self.0.add_update_auth_key(name, key)
+    }
+
+    /// Add Signer
     fn add_zone_signing_key(&mut self, signer: Signer) -> DnsSecResult<()> {
-        use trust_dns::rr::rdata::DNSSECRData;
-
-        // also add the key to the zone
-        let zone_ttl = self.minimum_ttl();
-        let dnskey = signer.key().to_dnskey(signer.algorithm())?;
-        let dnskey = Record::from_rdata(
-            self.origin.clone().into(),
-            zone_ttl,
-            RData::DNSSEC(DNSSECRData::DNSKEY(dnskey)),
-        );
-
-        // TODO: also generate the CDS and CDNSKEY
-        let serial = self.serial();
-        self.upsert(dnskey, serial);
-        self.secure_keys.push(signer);
-        Ok(())
+        self.0.add_zone_signing_key(signer)
     }
 
-    /// This will fail, the dnssec feature must be enabled
-    #[cfg(not(feature = "dnssec"))]
-    fn add_zone_signing_key(&mut self, _signer: Signer) -> DnsSecResult<()> {
-        Err("DNSSEC was not enabled during compilation.".into())
-    }
-
-    /// (Re)generates the nsec records, increments the serial number nad signs the zone
-    #[cfg(feature = "dnssec")]
+    /// Sign the zone for DNSSEC
     fn secure_zone(&mut self) -> DnsSecResult<()> {
-        // TODO: only call nsec_zone after adds/deletes
-        // needs to be called before incrementing the soa serial, to make sur IXFR works properly
-        self.nsec_zone();
-
-        // need to resign any records at the current serial number and bump the number.
-        // first bump the serial number on the SOA, so that it is resigned with the new serial.
-        self.increment_soa_serial();
-
-        // TODO: should we auto sign here? or maybe up a level...
-        self.sign_zone()
-    }
-
-    /// (Re)generates the nsec records, increments the serial number nad signs the zone
-    #[cfg(not(feature = "dnssec"))]
-    fn secure_zone(&mut self) -> DnsSecResult<()> {
-        Err("DNSSEC was not enabled during compilation.".into())
+        Authority::secure_zone(&mut self.0)
     }
 }
 
-// TODO: construct a battery of standard authority tests
 #[cfg(test)]
 mod tests {
     use super::*;
     use authority::ZoneType;
     use std::net::Ipv4Addr;
     use std::str::FromStr;
+    use trust_dns::rr::RData;
 
     #[test]
     fn test_load_zone() {
