+++ crates/server/src/store/in_memory/mod.rs
@@ -1,18 +1,12 @@
-// Copyright 2015-2018 Benjamin Fry <benjaminfry -@- me.com>
+// Copyright 2015-2019 Benjamin Fry <benjaminfry@me.com>
 //
 // Licensed under the Apache License, Version 2.0, <LICENSE-APACHE or
 // http://apache.org/licenses/LICENSE-2.0> or the MIT license <LICENSE-MIT or
 // http://opensource.org/licenses/MIT>, at your option. This file may not be
 // copied, modified, or distributed except according to those terms.
 
-//! All persistent store implementations
+//! Master file based serving with Dynamic DNS and journaling support
 
-mod config;
-pub mod file;
-pub mod forwarder;
-pub mod in_memory;
-pub mod sqlite;
+mod authority;
 
-// TODO: add a dynamic library option?
-
-pub use self::config::StoreConfig;
+pub use self::authority::InMemoryAuthority;
